backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
marshmallow==3.21.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.3
faker==24.0.0
gunicorn==21.2.0
APScheduler==3.10.4
anthropic>=0.40.0
pgvector>=0.3.0
voyageai>=0.3.0
